Make 'TenderlySolcConfig' depend on 'SolcConfig', then modify the 'libraries' type (#84)
In order to expand the solidity compiler configuration that verification can have on the TenderlySDK, we need to make `TenderlySolcConfig` depend on the `SolcConfig` because there are slight changes between the real solidity compiler configuration and what TenderlySDK sends to the backend server.

This PR makes `TenderlySolcConfig`'s settings depend on `SolcConfig`'s settings.